api/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
requests==2.31.0
twilio==8.5.0
pydantic==2.3.0
python-multipart==0.0.6
aiofiles==23.1.0
websockets==11.0.3
httpx==0.23.0
psycopg2-binary==2.9.9
